MFC: Minor cleanups to make some 64-bit platforms happier
- Bug #2106: use size_t for internal cbdata offset calculation
- Bug #2107: Unaligned memory access in pinger
- Make GCC happier on certain 64-bit platforms

+++ src/cbdata.c
@@ -122,7 +122,7 @@ cbdata_hash(const void *p, unsigned int mod)
 }
 
 #else
-#define OFFSET_OF(type, member) ((int)(char *)&((type *)0L)->member)
+#define OFFSET_OF(type, member) ((size_t)(char *)&((type *)0L)->member)
 #endif
 
 void
